Allow to not overwrite .gitconfig
git-svn-id: http://svn.shatow.net/shells@153 ea227b21-592c-dd11-b15d-00012992a58a

+++ update.sh
@@ -41,7 +41,7 @@ install -v dot.profile.common ~/.profile.common
 install -v dot.bash_profile ~/.bash_profile
 install -v dot.bashrc ~/.bashrc
 install -v dot.nanorc ~/.nanorc
-install -v dot.gitconfig ~/.gitconfig
+[ -f ~/.keep-gitconfig ] || install -v dot.gitconfig ~/.gitconfig
 install -v dot.supp ~/.supp
 install -v dot.valgrindrc ~/.valgrindrc
 install -v dot.bash_logout ~/.bash_logout
